requirements.txt
# Python 3.11

langchain
langchain-core
langchain-community
langchain-ollama
langchain-openai

chromadb
python-dotenv
requests
tqdm
numpy